refactor: wrap user responses
