+++ src/main/java/com/ahumadamob/fnanz/dto/response/ApiResponseErrorDto.java
@@ -1,4 +1,4 @@
-package com.ahumadamob.fnanz.error.dto;
+package com.ahumadamob.fnanz.dto.response;
 
 import java.time.Instant;
 import java.util.List;
